iaps-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
dnspython==2.4.2
email-validator==2.1.0
Werkzeug==3.0.1